BREAKING CHANGE: fix parsing of IPs with less than 4 bytes
Previous API was treating IPs with less than for bytes as IP with a
netmask of the size of the provided bytes (1=8, 2=16, 3=24) and was
interpreting the IP as if it was completed with 0s on the right side.

Propre IP parsing for these is to consider missing bytes as being 0s on
the left side.

Mask size is no longer infered by the number of bytes provided.